compra-entradas/requirements.txt
Flask==2.3.3
Flask-MySQL==1.5.2
PyMySQL==1.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
cryptography==41.0.4